Implements layout change on resizing and clicking on images to do actions.
Replaces the Grid with a FlowBox and puts it in a ScrolledWindow to
automatically change the layout as the width changes. Also puts
the image inside of a Button in order to be able to click it to
perform actions (later on to open more details about that particular
podcast).