Fixes bug, more details in the extended message.
If a download was attempted and the local file already existed,
it would cause an error and do nothing. Now at least when the
download is attempted the buttons update silently. There is still
a bug somewhere with the downloaded status of episodes either not
getting set correctly upon the library being loaded, or not getting
set in the database correctly when the episode is downloaded, or
inadvertently getting changed in the database during some other
operation. Further investigation into that is ongoing.