Improves starting playback.
There is a delay between hitting the play button and actually
starting playback. For some reason, checking the duration of an
episode sometimes results in a duration of 0. I fixed this by adding
a timeout that checks the duration and sets playback position and
starts playback when the duration is not 0, then stops repeating the
check. I shortened the timeout length and moved the play function
call into the check for duration != 0. This should result in a smoother
experience.