I think I fixed the erroneous re-init of episodes bug
Basically the INSERT OR REPLACE in write episode to database was
causing the last played epsiode to be moved to the bottom row of
the databse when its playback position (or anything really) was
edited. The update_feeds function would then see the last played
episode as the most recent episode added to the database. I fixed
it by changing REPLACE to UPDATE so it wouldn't move any rows
around when anything about an episode was edited.