Makes some changes necessary for flatpak to work correctly.
This unfortunately required removing the ability to change the
library location setting, I may be able to find a way
to reimplement this. In order for files to be downloaded while
running in the sandboxed flatpak, I needed the --talk-name=org.gtk.vfs
lines. I removed the check for the library location before figuring
this out so technically it could be reimplemented. I would want the
default location to be inside the sandbox by default anyway, and the
previous implementation wasn't really compatible with that anyway.
So some amount of work to fix that would've been necessary regardless.
For now the location is not configurable, but it likely will be later