Fixes bug that caused crash if podcast was deleted.
Somehow the old podcast episodes view for the deleted podcast hung
around and would appear after the podcast was deleted. So if you
clicked on another podcast it would switch to that page instead of
the one for the podcast you clicked on. This removes the deleted
podcast's episodes view when a podcast is deleted to prevent this.