Adds an ArtworkPopover to each EpisodeListItem.
Previously the only way to view the full description of an episode was to download it, play it, and click the CoverArt in the PlaybackBox. Not the best experience. This fixes that. I may be changing that name soon and replacing the ArtworkPopover with something more aptly named later, but for now this works just fine.