Adds option for EpisodeListItem to display CoverArt
New Episodes View and PodcastView can now use the same widget for their List/FlowBox children and tell it whether or not to display the coverart, rather than NewEpisodesView needing to manually add the coverart and EpisodeListItem to a box. Also, an eventual Queue can use the same Widget as well.