Refactors Downloads
There is now a `DownloadManager` that keeps track of a list of download objects, and `DownloadsPopover` (soon to be replaced with a window possibly) now keeps track of the list using a `bind_model` call. Additionally, the `DownloadsButton` now has a progress bar overlay and the `DownloadsPopover` no longer appears when a download starts.