Conversion to GTK4 (#24)
* Fixes icons for non-elementary distros

Need to see how these look on Elementary and figure out what I can do to make it
look decent on both

* Prefer use of built-in system icons.

* Gets app compiling with GTK4 and libadwaita

Lots of work still left to do to make it look and work properly

* Fixes margins on main window

* Fixes styling, ordering, etc. on PlaybackBox

* Implements ListModel binding for NewEpisodesView

* Implements bind_model for PodcastView episodes

* Fixes button changes when downloading or deleting

* Fixes styling of DownloadsPopover

* Improves styling on episode lists

* Update flatpak runtime version

* Adds some spacing to placeholder button.

* Fixes flatpak building

* Updates to libsoup-3.0

* Fixes bug with request_to_string

* Fixes icon used for downloads

* Switches style manager to Adw.StyleManager

* Fixes 32x32 icon (it was 32x34)

* Fix deprecation warning

* Fixes accelerator for play/pause button

* Fixes some warnings surrounding download details

* Prevents selecting episode list entries

* Adds trailing comma to meson.build dependencies

* Gets app back in line with Elementary guidelines

As a result, libadwaita is (for now) no longer a dependency. If I need a widget
from there it could come back in but I'm not expecing it to.

* Several layout changes to improve responsiveness

* Adds rounded corners to episode list items

* New icon

I think it's a little better but I am bad at art lol. Improvements are welcome

* Minor change to Coverart view

* Adds some padding to Coverart

* Refactors CoverArt to use GObject Style Construction

* Converts Episode to use GObject Style Construction

* Refactors Podcast to GObject Style Construction

* Should resolve potential bugs with new episodes

* Fixes bug with some default values

* Fixes more bugs surrounding new episodes

* First pass on async adding podcasts

* Remove some code duplication and fix warnings

* Re-adds podcast details and unsubscribe button.

Also has PodcastView use GOjbevt Style Construction.

* First pass on async loading episodes

* Cleans up several thread/async issues

* Implements bind_model for the podcasts list

* Async unsubscribe first pass

* More multithreading/async

* Styling/HIG compliance changes

* Removes verbatim strings from Library.vala

This is a DX thing, treesitter doesn't parse them correctly. Functionally identical to previous commit.

* MainWindow GObject Style Construction + Fix Warnings

* Library GObject Construction and fixed warnings

* More warnings fixes

* Fix all linter errors from vala-lint

vala-lint can be found at https://github.com/vala-lang/vala-lint

* Controller uses GObject style construction

* Asks for confirmation before unsubscribing

* Light reorganization

* Moves sorting to actual episodes list

Before it was on the ListBox but since I was using bind_model that was creating a warning. Instead I sort the actual episodes list.

* Fixes more warnings

* Adds ignore file for vala-lint

* Replaces ArtworkPopover with a Window

* Adds OverlayBar to show long-running background processes

* Redone PlaybackRatePopover to use a PopoverMenu

* Styling Cleanup, GObject Style Construction

* Adds option for EpisodeListItem to display CoverArt

New Episodes View and PodcastView can now use the same widget for their List/FlowBox children and tell it whether or not to display the coverart, rather than NewEpisodesView needing to manually add the coverart and EpisodeListItem to a box. Also, an eventual Queue can use the same Widget as well.

* Separated EpisodeList to a separate widget

Allows for better code sharing between PodcastView and NewEpisodesView and a future QueueView.

* Fixes regression in EpisodeListItem sizing

* Async loading of CoverArt images

* Refactors Downloads

There is now a `DownloadManager` that keeps track of a list of download objects, and `DownloadsPopover` (soon to be replaced with a window possibly) now keeps track of the list using a `bind_model` call. Additionally, the `DownloadsButton` now has a progress bar overlay and the `DownloadsPopover` no longer appears when a download starts.

* Fixes missed call to `cancelled` signal

* Improved algorithm for total DownloadManager percentage

It now has an accumulator that resets whenever all downloads have finished, and uses that as the "size" to divide by. This results in a smoother, less jerky progress bar filling. When all downloads have completed or been cancelled, the accumulator resets to 0.

* Replaces `DownloadsPopover` with `DownloadsWindow`.

* Colorscheme init cleanup

* Some cleanup on where download paths are set

more to come on this front later.

* Removes temp info logs added for troubleshooting

* Resolves some MPRIS related critical logs in console

* Fixes some Critical and Warning logs

* Fixes critical logs when deleting a podcast

* Fixes critical log when adding a podcast.

* Various bits of reorganization to improve startup experience.

Large amounts of podcasts were causing the app to take a long time to show the first window. This reorganizes and asyncs some of the order of operations on launch to display a window faster.

* Adds a loading page that displays on startup

Allows for the window to smoothly animate in before the podcasts load. Before they would start loading in before the window actually fades in causing the animation to freeze up and the window to just pop in.

* Adds spinner to episode download button after clicking.

* Moves Import to a menu and gives it a keyboard shortcut

* Reorganize app action code to Controller

* Reorganizing of actions to handle them in Controller

* Fix colorful buttons to use accent color

Also removes lots of unused css

* Store and restore window size/state

* Applies proper css classes to Seekbar